requirements_1.txt
python-telegram-bot==22.1
httpx==0.28.1
httpcore==1.0.9

